# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Read Station API",
        "description": "API for Read Station",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/invoice-enters": {
            "get": {
                "tags": [
                    "Admin / Invoice Enter"
                ],
                "summary": "Get all invoice enters",
                "description": "Lấy danh sách hóa đơn nhập",
                "operationId": "29f5e2762e8d5c683725233fe497fb19",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Số trang hiện tại",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Số lượng mục trên mỗi trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Từ khóa tìm kiếm",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Trạng thái",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive",
                                "deleted"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all suppliers successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/invoice-enters/{id}": {
            "get": {
                "tags": [
                    "Admin / Invoice Enter"
                ],
                "summary": "Get invoice enter by ID",
                "description": "Lấy thông tin hóa đơn nhập theo ID",
                "operationId": "9f61eb204b3dc9c3b864752a89ca9930",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID hóa đơn nhập",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get invoice enter successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/invoice-enters/create": {
            "post": {
                "tags": [
                    "Admin / Invoice Enter"
                ],
                "summary": "Create new invoice enter",
                "description": "Tạo mới một nhà cung cấp",
                "operationId": "7f70a01bb915e95a49e2f386dad19b10",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "invoice_code",
                                    "invoice_name",
                                    "total",
                                    "invoice_description",
                                    "supplier_id",
                                    "invoice_date",
                                    "status",
                                    "invoice_enter_detail"
                                ],
                                "properties": {
                                    "invoice_code": {
                                        "description": "Mã hóa đơn",
                                        "type": "string",
                                        "default": null,
                                        "nullable": true
                                    },
                                    "invoice_name": {
                                        "description": "Tên hóa đơn",
                                        "type": "string"
                                    },
                                    "total": {
                                        "description": "Tổng tiền",
                                        "type": "string"
                                    },
                                    "invoice_description": {
                                        "description": "Mô tả hóa đơn",
                                        "type": "string"
                                    },
                                    "supplier_id": {
                                        "description": "ID nhà cung cấp",
                                        "type": "integer"
                                    },
                                    "invoice_date": {
                                        "description": "Ngày hóa đơn",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "description": "Trạng thái",
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "active"
                                        ]
                                    },
                                    "invoice_enter_detail": {
                                        "description": "Chi tiết hóa đơn",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "book_detail_id",
                                                "book_price",
                                                "book_quantity"
                                            ],
                                            "properties": {
                                                "book_detail_id": {
                                                    "description": "ID chi tiết sách",
                                                    "type": "integer"
                                                },
                                                "book_price": {
                                                    "description": "Giá sách",
                                                    "type": "string"
                                                },
                                                "book_quantity": {
                                                    "description": "Số lượng sách",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create invoice enter successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/invoice-enters/update/{id}": {
            "put": {
                "tags": [
                    "Admin / Invoice Enter"
                ],
                "summary": "Update invoice enter by ID",
                "description": "Cập nhật thông tin hóa đơn nhập theo ID",
                "operationId": "94b4ca36f8d878bdc24477f874e0b832",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID hóa đơn nhập",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "invoice_name",
                                    "invoice_description",
                                    "status"
                                ],
                                "properties": {
                                    "invoice_name": {
                                        "description": "Tên hóa đơn",
                                        "type": "string"
                                    },
                                    "invoice_description": {
                                        "description": "Mô tả hóa đơn",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Trạng thái",
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "active",
                                            "canceled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update invoice enter successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}